+++ sdk/networkfunction/Azure.ResourceManager.NetworkFunction/tests/ScenarioTests/AzureTrafficCollectorTest.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
 using System;
@@ -186,7 +186,7 @@ namespace Azure.ResourceManager.NetworkFunction.Tests
             {
                 IngestionPolicy = new IngestionPolicyPropertiesFormat()
                 {
-                    IngestionType = IngestionType.IPFIX,
+                    IngestionType = IngestionType.Ipfix,
                     IngestionSources =
                     {
                         new IngestionSourcesPropertiesFormat()
@@ -200,7 +200,7 @@ namespace Azure.ResourceManager.NetworkFunction.Tests
                 {
                     new EmissionPoliciesPropertiesFormat()
                     {
-                        EmissionType = EmissionType.IPFIX,
+                        EmissionType = EmissionType.Ipfix,
                         EmissionDestinations =
                         {
                             new EmissionPolicyDestination()