+++ sdk/networkfunction/Azure.ResourceManager.NetworkFunction/tests/ScenarioTests/CollectorPolicyTest.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
 using System;
@@ -196,7 +196,7 @@ namespace Azure.ResourceManager.NetworkFunction.Tests
                 {
                     new EmissionPoliciesPropertiesFormat()
                     {
-                        EmissionType = EmissionType.IPFIX,
+                        EmissionType = EmissionType.Ipfix,
                         EmissionDestinations =
                         {
                             new EmissionPolicyDestination()